Added default action support for gallery's element. @TODO do the same for lists.
