+++ README.md
@@ -1,12 +1,10 @@
 # ChatFuel JSON API NodeJS library (TESTING)
-This library enables you to create the right JSON format for the Chatfuel JSON API.
-
-This library can be used ES5 with compatible Node versions. 
+This library enables you to create the right JSON format for the Chatfuel JSON API. It is lightweight (4.7k, 1.6k Gzipped) and is fully compatible with ES5 Node versions. 
 
 Chatfuel is an easy to use Messenger Bot creation platform.
 
 ## Chatfuel JSON API
-The Chatfuel JSON API is well documented on the Chatfuel.com website. It can be found [here](https://docs.chatfuel.com/api/json-api/json-api).
+The Chatfuel JSON API is well documented on the Chatfuel.com website. It can be found [inside Chatfuel's API doc](https://docs.chatfuel.com/api/json-api/json-api) and [Facebook Developer Plateform][https://developers.facebook.com/docs/messenger-platform/send-messages].
 
 ## Install
 
@@ -45,7 +43,7 @@ You can the send it using, with express for example:
 res.json(chatfuelled.toJson());
 ```
 
-### Methods
+### Methods (simple blocks)
 
 #### .addMessage (message)
 
@@ -81,6 +79,23 @@ chatfuelled.addImage("https://media.giphy.com/media/3oz8xPKZN7EwfcD0ys/giphy.gif
 | Attribute | Type   | Required | Description  |
 | --------- | ------ | -------- | ------------ |
 | url       | string | yes      | The file url |
+
+#### .addUserAttributes (name, value)
+
+Set user attributes without having the user to tap any button.
+
+| Attribute | Type   | Required | Description       |
+| --------- | ------ | -------- | ----------------- |
+| name      | string | yes      | Attribute's name  |
+| value     | string | yes      | Attribute's value |
+
+##### Example
+
+```js
+chatfuelled.addUserAttributes("age","30");
+```
+
+### Galleries
 
 #### .newGallery()
 
@@ -119,13 +134,13 @@ Messenger also supports specialised buttons:
 chatfuelled.sanitizeToButton("web_url", "https://www.google.ch/", "Go to Goole")
 ```
 
-#### .addElementToGallery (gallery_id, title, img_url, description, buttons...)
+#### .addElementToGallery (galleryID, title, img_url, description, buttons...)
 
-Add a new element to _gallery_id_. An element is by a title, an image, a description and buttons to request input from the user. Each element can have up to 3 buttons.
+Add a new element to _galleryID_. An element is by a title, an image, a description and buttons to request input from the user. Each element can have up to 3 buttons.
 
 | Attribute   | Type   | Required | Description                                                  |
 | ----------- | ------ | -------- | ------------------------------------------------------------ |
-| gallery_id  | int    | yes      | The gallery ID. Use _.newGallery()_ to get a new id          |
+| galleryID   | int    | yes      | The gallery ID. Use _.newGallery()_ to get a new id          |
 | title       | string | yes      | The title of the gallery                                     |
 | img_url     | string | yes      | The image url                                                |
 | description | string | yes      | Some description for the _subtitle_ tag                      |
@@ -133,10 +148,12 @@ Add a new element to _gallery_id_. An element is by a title, an image, a descrip
 | ...         |        | no       |                                                              |
 | button3     | JSON   | no       |                                                              |
 
+**Return** Element's ID (necessary for _addDefaultActionToElement_)
+
 ##### Example
 
 ```js
-chatfuelled.addElementToGallery(
+var elementID = chatfuelled.addElementToGallery(
   newGallery,
   "My first gallery",
   "https://www.jqueryscript.net/images/jQuery-Plugin-For-Stacked-Polaroid-Image-Gallery-Photopile.jpg",
@@ -145,6 +162,32 @@ chatfuelled.addElementToGallery(
   chatfuelled.sanitizeToButton("show_block", ["Bye 1", "Bye 2"], "bye")
 );
 ```
+
+#### .addDefaultActionToElement (galleryID, elementID, type,  url)
+
+**Optional** The default action executed when the template is tapped. 
+
+| Attribute            | Type   | Required | Description                                                  |
+| -------------------- | ------ | -------- | ------------------------------------------------------------ |
+| galleryID            | int    | yes      | The gallery ID. Use *.newGallery()* to get a new id          |
+| elementID            | int    | yes      | The element's ID. Use _.addElementToGallery()_ to get one    |
+| type                 | string | yes      | The default action type. Set to "web_view" by default.       |
+| url                  | string | yes      | The URL to which the user will be redirected when he/she tap over the gallery's element. |
+| messenger_extensions | bool   | no       | messenger_extensions                                         |
+| webview_height_ratio | string | no       | webview_height_ratio                                         |
+| fallback_url         | string | no       | fallback_url                                                 |
+
+##### Example
+
+```js
+chatfuelled.addDefaultActionToElement(
+  newGallery,
+  elementID,
+  "web_url",
+  "https://www.google.com/");
+```
+
+### Other methods
 
 #### .addButtons (message, buttons...)
 
@@ -166,21 +209,6 @@ chatfuelled.addButtons(
   chatfuelled.sanitizeToButton("web_url", "https://www.google.ch/maps", "Go to Maps"),
   chatfuelled.sanitizeToButton("show_block", ["Bye 1", "Bye 2"], "BYE 3")
 );
-```
-
-#### .addUserAttributes (name, value)
-
-Set user attributes without having the user to tap any button.
-
-| Attribute | Type   | Required | Description       |
-| --------- | ------ | -------- | ----------------- |
-| name      | string | yes      | Attribute's name  |
-| value     | string | yes      | Attribute's value |
-
-##### Example
-
-```js
-chatfuelled.addUserAttributes("age","30");
 ```
 
 ## TODO
